[proxima-direct-io-elasticsearch] module refactoring
