+++ direct/io-elasticsearch/src/main/java/cz/o2/proxima/direct/elasticsearch/ElasticsearchClientFactory.java
@@ -42,7 +42,7 @@ import org.elasticsearch.client.RestClient;
 import org.elasticsearch.client.RestClientBuilder;
 
 @Slf4j
-public class ElasticSearchClientFactory {
+public class ElasticsearchClientFactory {
 
   @Value
   @Builder
