+++ direct/io-elasticsearch/src/test/java/cz/o2/proxima/direct/elasticsearch/ElasticsearchClientFactoryTest.java
@@ -15,19 +15,23 @@
  */
 package cz.o2.proxima.direct.elasticsearch;
 
-import static cz.o2.proxima.storage.internal.AbstractDataAccessorFactory.*;
+import static cz.o2.proxima.direct.elasticsearch.ElasticsearchClientFactory.parseHosts;
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.net.URI;
+import org.apache.http.HttpHost;
 import org.junit.jupiter.api.Test;
 
-class ElasticSearchStorageTest {
+class ElasticsearchClientFactoryTest {
 
   @Test
-  public void testAccept() {
-    ElasticSearchStorage storage = new ElasticSearchStorage();
-    assertEquals(Accept.ACCEPT, storage.accepts(URI.create("elastic://asdas")));
-    assertEquals(Accept.ACCEPT, storage.accepts(URI.create("elasticsearch://asdas")));
-    assertEquals(Accept.REJECT, storage.accepts(URI.create("es://asdas")));
+  public void testParseHosts() {
+    HttpHost[] hosts = parseHosts("example.com:9093,example2.com", "http");
+    assertEquals(2, hosts.length);
+    assertEquals("example.com", hosts[0].getHostName());
+    assertEquals(9093, hosts[0].getPort());
+    assertEquals("http", hosts[0].getSchemeName());
+    assertEquals("example2.com", hosts[1].getHostName());
+    assertEquals(9200, hosts[1].getPort());
+    assertEquals("http", hosts[1].getSchemeName());
   }
 }
